hack easy-rsa

+++ roles/kubernetes/files/make-ca-cert.sh
@@ -96,7 +96,14 @@ service_ip=$(echo "${octets[*]}" | sed 's/ /./g')
 # Determine appropriete subject alt names
 sans="IP:${cert_ip},IP:${service_ip},DNS:kubernetes,DNS:kubernetes.default,DNS:kubernetes.default.svc,DNS:kubernetes.default.svc.${dns_domain},DNS:${master_name}"
 
-curl -sSL -O https://storage.googleapis.com/kubernetes-release/easy-rsa/easy-rsa.tar.gz
+#curl -sSL -O https://storage.googleapis.com/kubernetes-release/easy-rsa/easy-rsa.tar.gz
+rsa_tar_file="/tmp/easy-rsa.tar.gz"
+if [ -f "$rsa_tar_file" ]
+then
+  cp /tmp/easy-rsa.tar.gz easy-rsa.tar.gz
+else
+  curl -sSL -O https://storage.googleapis.com/kubernetes-release/easy-rsa/easy-rsa.tar.gz
+fi
 tar xzf easy-rsa.tar.gz
 cd easy-rsa-master/easyrsa3
